circle extend from closedshape donut extend from closedshape

+++ Q9/src/src/main/java/com/apical/ziv/q9/shapes/Point.java
@@ -19,10 +19,10 @@ public class Point implements Shape {
 	private static AtomicLong idGenerator = new AtomicLong();
 
 	private long id;
-	private float x;
-	private float y;
+	private double x;
+	private double y;
 
-	public Point(float x, float y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 		id = idGenerator.incrementAndGet();
@@ -47,19 +47,19 @@ public class Point implements Shape {
 		this.id = id;
 	}
 
-	public float getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(float x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public float getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
